// src/app/features/forms/enums/bussiness-rules-form.enum.ts
export enum BussinessRulesFunctionEnum {
  GENERATE = 'GENERATE',
  ADJUST = 'ADJUST',
  REVIEW = 'REVIEW',
  REVIEW_TEAM_LEADER = 'REVIEWTL',
  EDIT_OPENING_DATE = 'EDITOPENINGDATE',
  ADJUST_RETURNED = 'ADJUSTRETURNED',
  VIEW = 'VIEW',
  REVIEW_PROCUREMENT_FIDUCIARY_SPECIALIST = 'REVIEWPFE',
  PENDINGPUBLICATION = 'PENDINGPUBLICATION',
}